refactor: extract hyperopt-params test to separate file
